+++ backend/src/main/java/com/fabry/david/sprintform/repositories/TransactionRepository.java
@@ -6,10 +6,11 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
-public interface TransactionRepository extends MongoRepository<Transaction, Long> {
+public interface TransactionRepository extends MongoRepository<Transaction, BigInteger> {
     public List<Transaction> findAll();
 
     public Optional<Transaction> findById(Long id);
